estratto la funzione e messa dentro App. dataTrend adesso  è raccolto in un array che viene .map una volta che useState di dati è >0 . Per lo style potresti prende i loro nomi di dataTrend nel map, metterlo in paramentro class={x}  e aggiungerlo semplicemente al css con lo stile che vuoi
