+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/rossellamascia/wa/react/dashboard-covid-19/src/index.js":"1","/Users/rossellamascia/wa/react/dashboard-covid-19/src/App.js":"2","/Users/rossellamascia/wa/react/dashboard-covid-19/src/reportWebVitals.js":"3","/Users/rossellamascia/wa/react/dashboard-covid-19/src/serviceWorkerRegistration.js":"4","/Users/rossellamascia/wa/react/dashboard-covid-19/src/componets/Header.js":"5","/Users/rossellamascia/wa/react/dashboard-covid-19/src/componets/StartCard.js":"6","/Users/rossellamascia/wa/react/dashboard-covid-19/src/componets/RegionPositive.js":"7"},{"size":901,"mtime":1612204670631,"results":"8","hashOfConfig":"9"},{"size":1751,"mtime":1612442229882,"results":"10","hashOfConfig":"9"},{"size":364,"mtime":1612204237546,"results":"11","hashOfConfig":"9"},{"size":5064,"mtime":1612204237547,"results":"12","hashOfConfig":"9"},{"size":1041,"mtime":1612442512935,"results":"13","hashOfConfig":"9"},{"size":1720,"mtime":1612442110831,"results":"14","hashOfConfig":"9"},{"size":1628,"mtime":1612437249264,"results":"15","hashOfConfig":"9"},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},"16bkzn",{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"27","messages":"28","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},"/Users/rossellamascia/wa/react/dashboard-covid-19/src/index.js",[],["31","32"],"/Users/rossellamascia/wa/react/dashboard-covid-19/src/App.js",[],"/Users/rossellamascia/wa/react/dashboard-covid-19/src/reportWebVitals.js",[],"/Users/rossellamascia/wa/react/dashboard-covid-19/src/serviceWorkerRegistration.js",[],"/Users/rossellamascia/wa/react/dashboard-covid-19/src/componets/Header.js",[],"/Users/rossellamascia/wa/react/dashboard-covid-19/src/componets/StartCard.js",["33"],"/Users/rossellamascia/wa/react/dashboard-covid-19/src/componets/RegionPositive.js",[],{"ruleId":"34","replacedBy":"35"},{"ruleId":"36","replacedBy":"37"},{"ruleId":"38","severity":1,"message":"39","line":12,"column":21,"nodeType":"40","messageId":"41","endLine":12,"endColumn":27},"no-native-reassign",["42"],"no-negated-in-lhs",["43"],"no-unreachable","Unreachable code.","BreakStatement","unreachableCode","no-global-assign","no-unsafe-negation"]
+[{"C:\\Users\\noah\\Google Drive\\atom\\react-dashboard-covid19\\src\\index.js":"1","C:\\Users\\noah\\Google Drive\\atom\\react-dashboard-covid19\\src\\serviceWorkerRegistration.js":"2","C:\\Users\\noah\\Google Drive\\atom\\react-dashboard-covid19\\src\\App.js":"3","C:\\Users\\noah\\Google Drive\\atom\\react-dashboard-covid19\\src\\reportWebVitals.js":"4","C:\\Users\\noah\\Google Drive\\atom\\react-dashboard-covid19\\src\\componets\\Header.js":"5","C:\\Users\\noah\\Google Drive\\atom\\react-dashboard-covid19\\src\\componets\\StartCard.js":"6","C:\\Users\\noah\\Google Drive\\atom\\react-dashboard-covid19\\src\\componets\\RegionPositive.js":"7"},{"size":926,"mtime":1612537819595,"results":"8","hashOfConfig":"9"},{"size":5201,"mtime":1612537819596,"results":"10","hashOfConfig":"9"},{"size":3020,"mtime":1612539902973,"results":"11","hashOfConfig":"9"},{"size":377,"mtime":1612537819595,"results":"12","hashOfConfig":"9"},{"size":1070,"mtime":1612537819594,"results":"13","hashOfConfig":"9"},{"size":687,"mtime":1612539881774,"results":"14","hashOfConfig":"9"},{"size":1666,"mtime":1612537819594,"results":"15","hashOfConfig":"9"},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},"1esrvf3",{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},{"filePath":"21","messages":"22","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},"C:\\Users\\noah\\Google Drive\\atom\\react-dashboard-covid19\\src\\index.js",[],["31","32"],"C:\\Users\\noah\\Google Drive\\atom\\react-dashboard-covid19\\src\\serviceWorkerRegistration.js",[],"C:\\Users\\noah\\Google Drive\\atom\\react-dashboard-covid19\\src\\App.js",["33"],"C:\\Users\\noah\\Google Drive\\atom\\react-dashboard-covid19\\src\\reportWebVitals.js",[],"C:\\Users\\noah\\Google Drive\\atom\\react-dashboard-covid19\\src\\componets\\Header.js",[],"C:\\Users\\noah\\Google Drive\\atom\\react-dashboard-covid19\\src\\componets\\StartCard.js",[],"C:\\Users\\noah\\Google Drive\\atom\\react-dashboard-covid19\\src\\componets\\RegionPositive.js",[],{"ruleId":"34","replacedBy":"35"},{"ruleId":"36","replacedBy":"37"},{"ruleId":"38","severity":1,"message":"39","line":39,"column":17,"nodeType":"40","messageId":"41","endLine":39,"endColumn":23},"no-native-reassign",["42"],"no-negated-in-lhs",["43"],"no-unreachable","Unreachable code.","BreakStatement","unreachableCode","no-global-assign","no-unsafe-negation"]
